Display initial values for xm_Plnt, and output simulation time and BG_Output at end.
Looking at the way the next step is calculated in the loop, it appears
to me that Xaxis_time (likewise Xaxis_time_min) is the value at the
start of the step, while BG_Output is a value at the end of the step.
Accordingly I have added Ts (5 minutes) to Xaxis_time_min when
displaying it parallel to BG_Output.